Fix circular dependency in live slices
