+++ test/unit/slices/live/live-slice.test.ts
@@ -5,7 +5,8 @@ import { GameDetail, GameStatus } from '@app/models/game.js';
 import { LiveGame, LivePlayer, PeriodStatus, getPlayer } from '@app/models/live.js';
 import { PlayerStatus } from '@app/models/player.js';
 import { getGame as getGameCreator } from '@app/slices/game/game-slice.js';
-import { LiveState, actions, live, startGamePeriod } from '@app/slices/live/live-slice.js';
+import { live } from '@app/slices/live/composed-reducer.js';
+import { LiveState, actions, startGamePeriod } from '@app/slices/live/live-slice.js';
 import { RootState } from '@app/store.js';
 import { expect } from '@open-wc/testing';
 import sinon from 'sinon';
